+++ config/titan/scripts/adapt.groovy
@@ -23,4 +23,28 @@ if(! i) {
   mgmt.awaitGraphIndexStatus(graph, 'byIdent').status(SchemaStatus.ENABLED).call()
 } else { mgmt.commit() }
 
+// URL index
+mgmt = graph.openManagement()
+i = mgmt.getGraphIndex('byURL')
+if(! i) {
+  urlKey = mgmt.getPropertyKey('url')
+  urlKey = urlKey ? urlKey : mgmt.makePropertyKey('url').dataType(String.class).make()
+  mgmt.buildIndex('byURL', Vertex.class).addKey(urlKey, Mapping.STRING.asParameter()).buildMixedIndex('search')
+  mgmt.commit()
+  graph.tx().commit()
+
+  mgmt  = graph.openManagement()
+  urlKey = mgmt.getPropertyKey('url')
+  idx   = mgmt.getGraphIndex('byURL')
+  // Wait for index availability
+  if ( idx.getIndexStatus(urlKey).equals(SchemaStatus.INSTALLED) ) {
+    mgmt.commit()
+    mgmt.awaitGraphIndexStatus(graph, 'byURL').status(SchemaStatus.REGISTERED).call()
+  } else { mgmt.commit() }
+  mgmt  = graph.openManagement()
+  mgmt.updateIndex(mgmt.getGraphIndex('byURL'),SchemaAction.ENABLE_INDEX).get()
+  mgmt.commit()
+  mgmt.awaitGraphIndexStatus(graph, 'byURL').status(SchemaStatus.ENABLED).call()
+} else { mgmt.commit() }
+
 g = graph.traversal()